refactor(621): verplaats fromHeader-fuzzer naar fbs-common
De fuzzer op Identificatienummer.fromHeader stond in fbs-berichtensessiecache,
een toevallige consument. De geteste code leeft in fbs-common en wordt door beide
services gebruikt. Verplaatst naar libraries/fbs-common/src/test/.../fuzzing/ zodat
het vangnet co-located is met de code en blijft meelopen als de sessiecache wijzigt.

- Nieuw IdentificatienummerFromHeaderFuzzer in fbs-common (unieke bestandsnaam
  i.v.m. basename-afgeleide wrapper-namen in build.sh); invarianten ongewijzigd.
- Target + nu-ongebruikte imports verwijderd uit de sessiecache-DomainValidationFuzzer.
- jazzer-api test-dependency op fbs-common; libraries/fbs-common toegevoegd aan
  MODULES in .clusterfuzzlite/build.sh zodat de test-classes gekopieerd en de
  targets ontdekt worden.
